+++ packages/crypto_wallet_util/lib/forked_lib.dart
@@ -1,5 +1,3 @@
-library forked_lib;
-
 export 'src/forked_lib/psbt/psbt.dart';
 export 'src/forked_lib/bitcoin_flutter/bitcoin_flutter.dart';
 export 'src/forked_lib/tweetnacl-dart/tweetnacl.dart';