+++ packages/crypto_wallet_util/lib/src/config/config.dart
@@ -1,8 +1,4 @@
-library config;
-
-/**
- * This module stores various constants and predefined wallet configurations.
- */
+/// This module stores various constants and predefined wallet configurations.
 export '../config/chain/chains.dart';
 export '../config/constants/constants.dart';
 export '../config/constants/dictionary.dart';