+++ packages/crypto_wallet_util/lib/src/transaction/ckb/lib/ckb_lib.dart
@@ -1,4 +1,2 @@
-library ckb_lib;
-
 export './data_type.dart';
 export './serialize.dart';