+++ packages/crypto_wallet_util/lib/src/utils/bip39/bip39.dart
@@ -1,3 +1,1 @@
-library bip39;
-
 export 'src/bip39_base.dart';