+++ packages/crypto_wallet_util/lib/transaction.dart
@@ -1,6 +1,4 @@
 /// The cryptocurrencies that support transaction assembly are listed.
-library transaction;
-
 export 'src/transaction/alph/tx_signer.dart';
 export 'src/transaction/alph/tx_data.dart';
 export 'src/transaction/cosmos/tx_signer.dart';
@@ -55,6 +53,7 @@ export 'src/transaction/near/near_lib.dart';
 export 'src/transaction/xrp/xrp_lib.dart';
 export 'src/transaction/filecoin/fil_lib.dart';
 export 'src/transaction/sc/sc_lib.dart';
+export 'src/transaction/sc/sc_v2_semantics.dart';
 export 'src/transaction/sc/sc_wasm_bridge.dart' hide loadScWasm;
 export 'src/transaction/sc/sc_go_ffi_bridge.dart';
 export 'src/transaction/sc/sc_wasm_isolate_bridge.dart';